creating a server with express
